Use a VFS overlay to pass the incomplete Clang module to the compiler when building a mixed-language Swift module that generates a header.
The way we're currently building mixed-language modules that also generate headers causes problems with LLDB. What happens is this:

1. When building the Swift module, we have to pass in an "incomplete" C module that contains everything except the generated header (which hasn't been generated yet). This incomplete module gets serialized into the extra Clang flags of that module.
2. We take the generated header and compile a new C module, the "complete" one, and propagate that.
3. Downstream dependents now use and serialize the complete module in their extra Clang flags.
4. When LLDB loads a binary with both of these Swift modules, it combines the flags and tries to load two different modules (incomplete and complete) with the same name. This causes ASTContext creation to fail.

To resolve this, we can adopt the same technique that Xcode does; by using a VFS overlay with a specific name that the frontend explicitly ignores, we can make all the flags look the same even if they're pointing to different physical modules. Then, LLDB only sees one version of the module, and all is good.

PiperOrigin-RevId: 753165595