Start refactoring the Swift build APIs to remove the C++ toolchain as a dependency of the Swift toolchain.
Having the Swift toolchain propagate the C++ toolchain is wrong, because the execution platform decided for the `swift_toolchain` rule may not be the same one desired by individual `swift_*` rules that would have resolved that C++ toolchain differently. To address this, the Swift rules (and any custom rules that use the `swift_common` APIs) need to separately request the Swift and C++ toolchains.

Doing this directly (i.e., passing both toolchains separately) would make the Swift build APIs much harder to use, so instead we're adding new `{find,use}_all_toolchains` APIs that will request and return both in a convenient `struct`. The remaining public build APIs will be updated to take that `struct` instead of just the Swift toolchain provider, so those APIs can whichever provider—Swift or C++—it needs for a particular action.

This is CL 1 of _N_ to do this migration. Here, we update the internal APIs as needed and graft on new `toolchains` arguments to the public APIs. To keep existing clients working, we allow either `toolchains` or `swift_toolchain` to be passed, with the intention of removing the `swift_toolchain` parameter once clients have migrated.

PiperOrigin-RevId: 755893349