Ensure undeclared outputs are deleted when not tracked
In rules_swift < 3.x the .swiftsourceinfo files are uncondtionally written to the module path.In rules_swift >= 3.x these same files are no longer produced unless explicitly requested. When using non-sandboxed mode, previous builds will contain these files and cause build failures when Swift tries to use them, in order to work around this compatibility issue, we check the module path for the presence of .swiftsourceinfo files and if they are present but not requested, we remove them.

Testing:

- `bazel clean --expunge`
- `git checkout 2.8.2`
- `bazel build //examples/... --spawn_strategy=local` (pass)
- `git checkout master`
- `bazel build //examples/... --spawn_strategy=local` (failure)
- `git checkout <this-branch>`
- `bazel build //examples/... --spawn_strategy=local` (pass)