Unconditionally add the `testing` macro plugin subdirectory to compile actions.
Xcode 26.0 beta 5 has isolated this plugin to its own subdirectory and Xcode only passes it to the compiler when building test targets (previous versions had a symlink to make the default path work, but beta 5 removed that symlink). We don't want to make that distinction in our rules, so it's fine to just unconditionally pass it.

PiperOrigin-RevId: 795560788