Add a `feature_configuration` argument to `derive_swift_module_name`.
This is currently unused, but will be used later to detect the presence
of a feature that migrates modules from the current naming to one based
on Bazel target labels.

Naturally, this cannot be used by macros because they will not have
access to a feature configuration. Those will need to be migrated
separately, once all clients are on Swift 6.2/Xcode 26. The migration
will be "don't call the function" anymore, because there will be nothing
to derive; the module name will be the label verbatim. Likewise,
eventually this function can disappear in the rules themselves as well,
once the migration is complete.

(cherry picked from commit 596c252e521e9549a2b8940605305eb7a3931289)