Use apple constraints over config_setting_group in select
The intent of @apple_support//configs:apple is to provide a way for
users to select on whether the current platform is an apple one or not.

It does that by iterating over APPLE_PLATFORMS_CONSTRAINTS, and checking
whether the "cpus" is one of the apple known CPUs.
This works in the general case, but in the exceptional case where the
user wants to define custom apple platforms, it is a problem.

So in this patchset we're switching to using the constraint directly
over the config_setting_group(). It should be a functional no-op for
the general case, but it provides more flexibility if case the user
needs a higher degree of customization.