Fix layering check output for modules that are aliased.
The `-emit-imported-modules` compiler mode will resolve any aliases that it knows about, so if an alias is used to rename `Foo` to `RenamedFoo` (where clients are expected to still `import Foo` in source), the imported modules list will contain `RenamedFoo`. We map this back to the original name so that the message makes sense for the user reading it.

PiperOrigin-RevId: 800435022