Hoist the `original_module_name` field out of the module context's Swift-specific substructure into the main module context.
This was in the Swift substructure because module aliases originally only supported Swift modules. That's not _entirely_ true though; aliases can still be used for Clang modules. It won't impact the ABI (because C symbols don't contain their module name) but it still allows the module to be imported with a different name than what is written in the modulemap file. This will be used to support Bazel-labels-as-module-names.

PiperOrigin-RevId: 800467745