feat(toolchain): add embedded OS support in swift toolchain
When building for an embedded platform, we don't want to automatically
add the linux specific option to the linking command line. Specifically,
linking swiftrt.o references symbols that don't exist in that mode.

So this change adds support for a new `ctx.attr.os` value called
`none`. If selected in the toolchain, no OS specific linking options
will be added by default.

After this change, we are able to compile code in embedded mode. Trying
it out is a little dicey at this point, but it can be done as follow.
Pay attention to the various caveats in that process, more fixes are
required to make its use easier and more general. For the brave though,
here are some instructions:
* build on Linux - as rules_swift current assumes that whenever we're
  building on MacOS, we must be using an Xcode toolchain. Which isn't
  true when using Swift Embedded.
* download the Swift precompiled toolchain for your favorite Linux
  distribution from here:
  https://www.swift.org/install/linux/amazonlinux/2
  It's important to not use the swiftly toolchain, as rules_cc will
  resolve it incorrectly when trying to detect the compiler type.
* extract your toolchain and update your environment to have its bin
  folder in your PATH
* Run swift toolchain autoconfiguration by running:
  $ bazel build @build_bazel_rules_swift_local_config//:linux-toolchain
* Replace `os = "linux"` by `os = "none"` in
  bazel-rules_swift/external/+non_module_deps+build_bazel_rules_swift_local_config/BUILD
  Of course, in a real project, one would probably create a dedicated
  toolchain and associated platforms.
* Now build and run the hello_world example with the embedded feature:
  bazel run --features=swift.enable_embedded \
    //examples/xplatform/hello_world

At this point, the hello_world executable should run normally, but if
you look at its required runtime libraries, it shouldn't need any of the
standard swift libraries:
$ readelf -d bazel-bin/examples/xplatform/hello_world/hello_world|grep NEEDED
 0x0000000000000001 (NEEDED)             Shared library: [libc.so.6]