feat(toolchain): add missing action types to toolchain configs
The -target option should be added to the same actions, regardless of
whether we're declaring an Xcode toolchain or a standalone toolchain.
This commit brings the Linux actions up-to-date with the Xcode ones.

While we're at it, we're also adding -target to the MODULEWRAP action
to both Xcode and non-Xcode actions. It looks like it's ok to omit it
if the target arch is the same as the host arch, but if they're
different swiftc will return an error.