Add rpaths for the toolchain's compatibility back-deployment librarie (#1589)
This is necessary to ensure that these binaries run correctly when using
APIs like `Span` that are back-deployed on older OSes. Since the
`swift_*` rules don't do bundling of their own, we simply point to the
libraries in the toolchain itself (using the `xcode-select`-dependent
symlinks, as we were already doing for tests). To actually distribute
such a binary, it's the responsibility of the user to include the
required dylibs with it (or, better, use a rule from rules_apple to
perform the bundling).

I've explicitly ignored the swift-5.0 and swift-5.5 directory since we
have no plans to deploy anything from these rules to such old OSes.

Cherry pick:
[2f5aa6e6a3dae09293c9b35ff3cc5795fef279e7](https://github.com/bazelbuild/rules_swift/commit/2f5aa6e6a3dae09293c9b35ff3cc5795fef279e7)

---------

Co-authored-by: Tony Allevato <allevato@google.com>