Resolve all the current `buildifier` `lint` issues.
For the protos aspect, it seems it isn't smart enough to realize the two `if`s paths
are the same and thus things are always set. Avoid the false warnings by moving
things around a little so there is only one `if` path and ensure things are always
set that way instead.

PiperOrigin-RevId: 833982216