Add swift-embedded example
This will also serve as a test to make sure swift-embedded works
properly, along with the standalone toolchain extension.

The example is derived from the swift embedded example project here:
https://github.com/swiftlang/swift-embedded-examples

To try it, do the following:
$ cd examples/toolchain/embedded
$ bazelisk build --platforms=//:aarch64 //rpi-4b-blink:Application.bin

Note that at this stage, because this embedded code is making use of
SPM, and because SPM bazel rules currently tries to run
`swift package update` in its repository rules, having a swift
executable in the PATH is still a requirement.