refactor(toolchain): generalize swift_executable attr into swift_tools
When using `swift_executable` option to swift_toolchain(), individual
swift tools will be called using --driver-mode. This works when
switching between swift and swiftc, but tools such as
swift-autolink-extract and swift-symbolgraph-extract are separate and
aren't recognized by --driver-mode. Therefore, these actions
consistently fail in that scenario.

As we're trying to add standalone hermetic swift toolchain support, we
must have a way to include these tools in actions' input roots, so we're
replacing the `swift_executable` toolchain parameter with a
`swift_tools` one, which can be used to specify what binary to use for
each particular action, as bazel labels.

`swift_executable` is still available for backward compatibility, but it
will only used if `swift_tools` is not set.