fix(swift): disable -file_prefix_map feature on non-xcode toolchains
The file_prefix_map feature only works when DEVELOPER_DIR is set,
which is only available in an Xcode toolchain. As we're removing the
Apple macros in the bazel substitution code, we need to make sure
unexpected variables don't show-up in configuration where they're not
supported. We're doing so through the introduction of a new private
feature called `swift._supports_developer_dir`, that will be enabled
only on Xcode toolchains. Individual ActionConfigInfo can then check
that feature if they use the __BAZEL_XCODE_DEVELOPER_DIR__ string.