feat(tools): add utility to generate swift toolchain dictionaries
Generating the toolchain dictionary currently requires downloading
toolchains one by one and calculating their checksum, which is tedious
and time consuming.

We're adding a utility here that will do that automatically. The tool
supports caching downloaded archives and can fetch release information
from swift.org's API to generate the checksums dictionary format used
by the build system.

While we're at it, we're running that utility for the 2 latest releases
(6.2.2 and 6.2.3) and adding them to the dict.

Because that dictionary is probably gonna keep growing over time, we're
also moving it in its own standalone file (swift_releases.bzl).